Added soil moisture to ECS grid system

+++ Assets/Scripts/GridSystem/TestGrid.cs
@@ -1,8 +1,6 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
-public class TestGrid 
+public class TestGrid
 {
     private int width;
     private int height;
@@ -10,7 +8,7 @@ public class TestGrid
     private Vector3 originPosition;
 
 
-    private int[,] gridArray;
+    private float[,] gridArray;
     private TextMesh[,] debugTextArray;
 
 
@@ -21,7 +19,7 @@ public class TestGrid
         this.cellSize = cellSize;
         this.originPosition = originPosition;
 
-        gridArray = new int[width, height];
+        gridArray = new float[width, height];
         debugTextArray = new TextMesh[width, height];
 
 
@@ -33,7 +31,7 @@ public class TestGrid
             for (int z = 0; z < gridArray.GetLength(1); z++) // height
             {
                 //Debug.Log(x+", "+y);
-                debugTextArray[x, z] = GridHelper.CreateWorldText(gridArray[x, z].ToString(),null, GetWorldPosition(x, z) + new Vector3(cellSize, 0, cellSize) * 0.5f, 20, Color.white, TextAnchor.MiddleCenter);
+                debugTextArray[x, z] = GridHelper.CreateWorldText(gridArray[x, z].ToString(), null, GetWorldPosition(x, z) + new Vector3(cellSize, 0, cellSize) * 0.5f, 20, Color.white, TextAnchor.MiddleCenter);
                 Debug.DrawLine(GetWorldPosition(x, z), GetWorldPosition(x, z + 1), Color.white, 100f);
                 Debug.DrawLine(GetWorldPosition(x, z), GetWorldPosition(x + 1, z), Color.white, 100f);
 
@@ -62,14 +60,14 @@ public class TestGrid
     }
 
     // set value of cell in the array
-    public void SetValue(int x, int z, int value)
+    public void SetValue(int x, int z, float value)
     {
         if (x < 0 || z < 0 || x >= width || z >= height) // validation
             return;
 
         gridArray[x, z] = value;
 
-        if(gridArray[x, z] == 0)
+        if (gridArray[x, z] == 0)
         {
             debugTextArray[x, z].text = "";
         }
@@ -80,14 +78,14 @@ public class TestGrid
     }
 
 
-    public void SetValue(Vector3 worldPosition, int Value)
+    public void SetValue(Vector3 worldPosition, float Value)
     {
         int x, z;
-        GetXZ(worldPosition, out x, out z); 
+        GetXZ(worldPosition, out x, out z);
         SetValue(x, z, Value);
     }
 
-    public int GetValue(int x, int z)
+    public float GetValue(int x, int z)
     {
         if (x < 0 || z < 0 || x >= width || z >= height) // validation
             return 0;
@@ -97,7 +95,7 @@ public class TestGrid
 
 
 
-    public int GetValue(Vector3 worldPosition)
+    public float GetValue(Vector3 worldPosition)
     {
         int x, z;
         GetXZ(worldPosition, out x, out z);
@@ -120,7 +118,7 @@ public class TestGrid
 
 
 
-    public void Increment(int x, int z, int value = 1)
+    public void Increment(int x, int z, float value = 1)
     {
         if (x < 0 || z < 0 || x >= width || z >= height) // validation
             return;
